+++ src/main/java/friendlybot/command/Command.java
@@ -1,3 +1,9 @@
+package friendlybot.command;
+
+import friendlybot.Storage;
+import friendlybot.Ui;
+import friendlybot.task.TaskList;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
