+++ src/main/java/friendlybot/command/MarkCommand.java
@@ -1,3 +1,10 @@
+package friendlybot.command;
+
+import friendlybot.Storage;
+import friendlybot.Ui;
+import friendlybot.task.TaskList;
+import friendlybot.task.Task;
+
 public class MarkCommand extends Command {
     private boolean mark;
     private int taskNumber;
