+++ src/main/java/friendlybot/task/TaskList.java
@@ -1,3 +1,7 @@
+package friendlybot.task;
+
+import friendlybot.Ui;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
