+++ text-ui-test/runtest.bat
@@ -5,17 +5,19 @@ if not exist ..\bin mkdir ..\bin
 
 REM delete output from previous run
 if exist ACTUAL.TXT del ACTUAL.TXT
+if exist data rmdir /s /q data
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\*.java
+javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\friendlybot\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
+    pause
     exit /b 1
 )
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin FriendlyBot < input.txt > ACTUAL.TXT
+java -classpath ..\bin friendlybot.FriendlyBot < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
